Move Dry::Container to Dry::Core namespace
