--- lib/dry/container/config.rb
@@ -1,32 +1,0 @@
-# frozen_string_literal: true
-
-module Dry
-  class Container
-    # @api public
-    class Config
-      DEFAULT_NAMESPACE_SEPARATOR = "."
-      DEFAULT_RESOLVER = Resolver.new
-      DEFAULT_REGISTRY = Registry.new
-
-      # @api public
-      attr_accessor :namespace_separator
-
-      # @api public
-      attr_accessor :resolver
-
-      # @api public
-      attr_accessor :registry
-
-      # @api private
-      def initialize(
-        namespace_separator: DEFAULT_NAMESPACE_SEPARATOR,
-        resolver: DEFAULT_RESOLVER,
-        registry: DEFAULT_REGISTRY
-      )
-        @namespace_separator = namespace_separator
-        @resolver = resolver
-        @registry = registry
-      end
-    end
-  end
-end
