+++ spec/dry/container_spec.rb
@@ -1,14 +1,14 @@
 # frozen_string_literal: true
 
-RSpec.describe Dry::Container do
-  let(:klass) { Dry::Container }
+RSpec.describe Dry::Core::Container do
+  let(:klass) { Dry::Core::Container }
   let(:container) { klass.new }
 
   it_behaves_like "a container"
 
   describe "inheritance" do
     it "sets up a container for a child class" do
-      parent = Class.new { extend Dry::Container::Mixin }
+      parent = Class.new { extend Dry::Core::Container::Mixin }
       child = Class.new(parent)
 
       parent.register(:foo, Object.new)
